+++ lib/FinkExtension.php
@@ -3,10 +3,13 @@
 namespace DTL\Extension\Fink;
 
 use DTL\Extension\Fink\Console\Command\CrawlCommand;
-use DTL\Extension\Fink\Console\Display\ConcatenatingDisplay;
+use DTL\Extension\Fink\Console\DisplayBuilder;
+use DTL\Extension\Fink\Console\DisplayRegistry;
+use DTL\Extension\Fink\Console\Display\MemoryDisplay;
 use DTL\Extension\Fink\Console\Display\RateDisplay;
 use DTL\Extension\Fink\Console\Display\StatusLineDisplay;
 use DTL\Extension\Fink\Console\Display\ReportListDisplay;
+use DTL\Extension\Fink\Console\Display\UptimeDisplay;
 use DTL\Extension\Fink\Model\DispatcherBuilderFactory;
 use Phpactor\Container\Container;
 use Phpactor\Container\ContainerBuilder;
@@ -26,15 +29,23 @@ class FinkExtension implements Extension
         $container->register('fink.console.command.crawl', function (Container $container) {
             return new CrawlCommand(
                 $container->get(self::SERVICE_DISPATCHER_BUILDER_FACTORY),
-                $container->get('fink.console.display')
+                $container->get('fink.console.display_builder')
             );
         }, [ ConsoleExtension::TAG_COMMAND => [ 'name' => 'crawl' ]]);
 
-        $container->register('fink.console.display', function (Container $container) {
-            return new ConcatenatingDisplay([
-                new ReportListDisplay(),
-                new StatusLineDisplay(),
-                new RateDisplay(),
+        $container->register('fink.console.display_registry', function (Container $container) {
+            return new DisplayRegistry([
+                'urllist' => new ReportListDisplay(),
+                'status' => new StatusLineDisplay(),
+                'rate' => new RateDisplay(),
+                'memory' => new MemoryDisplay(),
+                'uptime' => new UptimeDisplay(),
+            ]);
+        });
+
+        $container->register('fink.console.display_builder', function (Container $container) {
+            return new DisplayBuilder($container->get('fink.console.display_registry'), [
+                'urllist', 'status', 'rate', 'uptime'
             ]);
         });